Maven

+++ Hotel1/repository-api/src/main/java/com/senlainc/git_courses/java_training/petushok_valiantsin/api/repository/IOrderDao.java
@@ -7,11 +7,11 @@ import com.senlainc.git_courses.java_training.petushok_valiantsin.model.Room;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface IOrderDao extends ICommonDao<Order, Integer> {
+public interface IOrderDao extends ICommonDao<Order, Long> {
 
     void update(Order order, Attendance attendance);
 
-    List<Room> readLastRoom(Integer index);
+    List<Room> readLastRoom(Long index);
 
     List<Room> readAfterDate(LocalDate date);
 }
