+++ Hotel1/repository-api/src/main/java/com/senlainc/git_courses/java_training/petushok_valiantsin/api/repository/IRoomDao.java
@@ -4,7 +4,7 @@ import com.senlainc.git_courses.java_training.petushok_valiantsin.model.Room;
 
 import java.util.List;
 
-public interface IRoomDao extends ICommonDao<Room, Integer> {
+public interface IRoomDao extends ICommonDao<Room, Long> {
 
     List<Room> readAllFree();
